+++ app/Http/Requests/ContractRequest.php
@@ -20,8 +20,13 @@ class ContractRequest extends FormRequest
     {
         return [
             'name' => 'required',
+            'value' => 'required',
             'start_day' => 'required',
-            'contract_type' => 'required'
+            'finish_day' => 'required',
+            'contract_type' => 'required',
+            'user_id' => 'required',
+            'customer_id' => 'required',
+            'contact_id' => 'required'
         ];
     }
 
@@ -29,9 +34,13 @@ class ContractRequest extends FormRequest
     {
         return [
             'name.required' => 'Tên hợp đồng không được để trống!',
-            'customer.required'=> 'Hãy chọn khách hàng!',
             'start_day.required' => 'Hãy chọn ngày bắt đầu hợp đồng!',
-            'contract_type.required' => 'Hãy chọn loại hợp đồng!'
+            'finish_day.required' => 'Hãy chọn ngày kết thúc hợp đồng!',
+            'value.required' => 'Giá trị hợp đồng không thể trống!',
+            'contract_type.required' => 'Hãy chọn loại hợp đồng!',
+            'user_id.required' => 'Hãy chọn chủ sở hữu!',
+            'customer_id.required' => 'Hãy chọn khách hàng bằng cách nhấn nút "Chọn Khách Hàng" phía trên!',
+            'contact_id.required' => 'Hãy chọn liên hệ bằng cách nhấn nút "Chọn Liên Hệ" phía trên!'
         ];
     }
 }
