+++ database/seeders/DatabaseSeeder.php
@@ -14,6 +14,15 @@ class DatabaseSeeder extends Seeder
      */
     public function run(): void
     {
+        // DB::table('users')->insert([
+        //     'name' => 'Hà Trung Nghĩa',
+        //     'email' => 'nghiauser@gmail.com',
+        //     'password' => bcrypt('123456789'),
+        //     'phone' => '0339557275',
+        //     'status' => 1, // Type USER
+        //     'address' => 'An Giang',
+        //     'created_at' => Carbon::now()
+        // ]);
         DB::table('users')->insert([
             'name' => 'Lê Trường',
             'email' => 'truong123@gmail.com',
@@ -38,6 +47,6 @@ class DatabaseSeeder extends Seeder
 
         //     'created_at' => Carbon::now()
         // ]);
-        
+
     }
 }
