+++ resources/views/frontend/layouts/business_transaction.blade.php
@@ -120,7 +120,7 @@
                     <!-- Profile Info -->
                     <div class="col-sm-6 clearfix">
                         <div class="user-profile pull-right">
-                            {{-- <img class="avatar user-thumb" src="{{ pare_url_file(Auth::user()->avatar) }}" onerror="this.src='/assets/images/author/avatar.png'"alt="avatar"> --}}
+                            <img class="avatar user-thumb" src="{{ pare_url_file(Auth::user()->avatar) }}" onerror="this.src='/assets/images/author/avatar.png'"alt="avatar">
                             <h4 class="user-name dropdown-toggle" data-toggle="dropdown">{{ Auth::user()->name ?? "[N\A]" }}<i class="fa fa-angle-down"></i></h4>
                             <div class="dropdown-menu">
                                 <a class="dropdown-item profile-option" href="#">Hướng dẫn sử dụng</a>
