+++ resources/views/frontend/layouts/customer.blade.php
@@ -315,7 +315,7 @@
                     <div class="col-sm-6 clearfix">
                         <div class="user-profile pull-right">
 
-                            {{-- <img class="avatar user-thumb" src="{{ Chatify::getUserWithAvatar(Auth::user())->avatar }}" onerror="this.src='/assets/images/author/avatar.png'" alt="avatar"> --}}
+                            <img class="avatar user-thumb" src="{{ Chatify::getUserWithAvatar(Auth::user())->avatar }}" onerror="this.src='/assets/images/author/avatar.png'" alt="avatar">
                             <h4 class="user-name dropdown-toggle" data-toggle="dropdown">{{ Auth::user()->name ?? "[N\A]" }}<i class="fa fa-angle-down"></i></h4>
 
                             <img class="avatar user-thumb"
