+++ resources/views/frontend/transaction/form_create.blade.php
@@ -33,7 +33,7 @@
                 @enderror
             </div>
             <div class="form-group">
-                <label for="" class="col-form-label input-label">Liên hệ:</label>
+                <label for="" class="col-form-label input-label">Liên hệ<code>*</code>:</label>
                 <select name="contact_id" class="custom-select custom-select-height" id="">
                     <option value="">----Chọn liên hệ----</option>
                     @foreach ($contacts ?? [] as $item)
@@ -72,6 +72,22 @@
             <input type="hidden" name="selected_finish_day" id="selected_finish_day">
         </div>
         <div class="col-sm-6">
+            <div class="row">
+                <div class="col-6 form-group">
+                    <label for="user_id" class="col-form-label input-label">Mã phụ trách:</label>
+                    <input type="text" name="user_id" class="form-control" value="{{ old('user_id', Auth::user()->id ?? '') }}" id="user_id" readonly>
+                    @error('user_id')
+                        <small id="" class="form-text text-danger">{{ $errors->first('user_id') }}</small>
+                    @enderror
+                </div>
+                <div class="col-6 form-group">
+                    <label for="user_name" class="col-form-label input-label">Người phụ trách:</label>
+                    <input type="text" name="user_name" class="form-control" value="{{ old('user_name', Auth::user()->name ?? '') }}" id="user_name" readonly>
+                    @error('user_name')
+                        <small id="" class="form-text text-danger">{{ $errors->first('user_name') }}</small>
+                    @enderror
+                </div>
+            </div>
             <div class="form-group">
                 <label for="choose_transaction" class="col-form-label input-label">Loại giao dịch<span style="color: red">*</span>:</label>
                 <select name="transaction_type" class="custom-select custom-select-height" id="choose_transaction">
